refactor(db): Remove database backup & restore logic
This commit removes the database backup and restore logic from the `run.sh` script.

The database backup and restore logic was originally added as a way to protect against data loss in the event of a database corruption. However, this logic is no longer necessary because Litestream now provides built-in backup and restore functionality.

Removing this logic simplifies the `run.sh` script and makes it more reliable.